Adding spec tests
