+++ Cargo.toml
@@ -8,10 +8,16 @@ license = "MIT"
 description = "An implementation of a lockfile used in Deno"
 repository = "https://github.com/denoland/deno_lockfile"
 
+[[test]]
+name = "specs"
+path = "tests/spec_test.rs"
+harness = false
+
 [dependencies]
 serde = { version = "1.0.149", features = ["derive"] }
 serde_json = "1.0.85"
 thiserror = "1.0.40"
 
 [dev-dependencies]
+file_test_runner = "0.7.0"
 pretty_assertions = "1.4.0"
